% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagger-walkthrough-app.R
\name{run_fluree_tagger_app}
\alias{run_fluree_tagger_app}
\title{Launch the Fluree tagging walkthrough}
\usage{
run_fluree_tagger_app(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
A Shiny app object, invisibly when launched.
}
\description{
Launch the Fluree tagging walkthrough
}
